+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `Typography`: 
+  - Remplace `tiny, regular, default, big` by `small, medium, large` prop size values.
 - Fix: Helpers: Add `checkDeprecatedSizes` utility.
 
 ## [9.3.1] - 2022-03-29